+++ awtk-widget-chart-view/src/chart_view/line_series.c
@@ -20,7 +20,6 @@
  */
 
 #include "tkc/utils.h"
-#include "chart_animator.h"
 #include "axis.h"
 #include "series_p.h"
 #include "line_series.h"
@@ -136,24 +135,26 @@ static ret_t line_series_set_value(widget_t* widget, const char* value) {
   const char* token = NULL;
   tokenizer_t tokenizer;
   float_t v;
-  fifo_t* fifo;
+  object_t* fifo;
+  uint32_t capacity;
   series_t* series = SERIES(widget);
   return_value_if_fail(series != NULL && value != NULL, RET_BAD_PARAMS);
 
-  fifo = fifo_create(series->capacity, series->unit_size, NULL, NULL);
+  capacity = widget_get_prop_int(widget, SERIES_PROP_CAPACITY, 0);
+  fifo = series_fifo_default_create(capacity, sizeof(float_t));
   return_value_if_fail(fifo != NULL, RET_OOM);
 
   tokenizer_init(&tokenizer, value, strlen(value), ",");
 
-  while (tokenizer_has_more(&tokenizer) && fifo->size < fifo->capacity) {
+  while (tokenizer_has_more(&tokenizer) && SERIES_FIFO_GET_SIZE(fifo) < capacity) {
     token = tokenizer_next(&tokenizer);
     v = tk_atof(token);
-    fifo_push(fifo, &v);
+    series_fifo_push(fifo, &v);
   }
 
-  series_set(widget, 0, fifo->buffer, fifo->size);
+  series_set(widget, 0, SERIES_FIFO_DEFAULT(fifo)->buffer, SERIES_FIFO_GET_SIZE(fifo));
 
-  fifo_destroy(fifo);
+  OBJECT_UNREF(fifo);
   tokenizer_deinit(&tokenizer);
 
   return RET_OK;
@@ -177,13 +178,13 @@ ret_t line_series_on_destroy(widget_t* widget) {
 
   TKMEM_FREE(series->series_axis);
   TKMEM_FREE(series->value_axis);
-  fifo_destroy(series->base.fifo);
+  OBJECT_UNREF(series->base.fifo);
 
   return RET_OK;
 }
 
 ret_t line_series_draw_one_series(widget_t* widget, canvas_t* c, float_t ox, float_t oy,
-                                  fifo_t* fifo, uint32_t index, uint32_t size, rect_t* clip_rect,
+                                  object_t* fifo, uint32_t index, uint32_t size, rect_t* clip_rect,
                                   series_p_draw_line_t draw_line,
                                   series_p_draw_line_area_t draw_area,
                                   series_p_draw_smooth_line_t draw_smooth_line,
@@ -236,6 +237,7 @@ ret_t line_series_draw_one_series(widget_t* widget, canvas_t* c, float_t ox, flo
   if (series->symbol.show) {
     widget_t* axis = widget_get_prop_pointer(widget, SERIES_PROP_SERIES_AXIS);
     float_t range = axis_get_range(axis, TRUE);
+    uint32_t fifo_size = SERIES_FIFO_GET_SIZE(fifo);
 
     // 默认symbol的border_with不超过symbol.size
     if (vertical) {
@@ -247,7 +249,7 @@ ret_t line_series_draw_one_series(widget_t* widget, canvas_t* c, float_t ox, flo
     }
 
     // caover类型时，旧波形减少一个sysmbol, 避免symbol的间距很小时，显示残余
-    if (fifo->size > range && index + size < fifo->size) {
+    if (fifo_size > range && index + size < fifo_size) {
       index++;
       size--;
     }
@@ -262,7 +264,8 @@ ret_t line_series_draw_one_series(widget_t* widget, canvas_t* c, float_t ox, flo
 }
 
 static ret_t line_series_on_paint(widget_t* widget, canvas_t* c, float_t ox, float_t oy,
-                                  fifo_t* fifo, uint32_t index, uint32_t size, rect_t* clip_rect) {
+                                  object_t* fifo, uint32_t index, uint32_t size,
+                                  rect_t* clip_rect) {
   return line_series_draw_one_series(
       widget, c, ox, oy, fifo, index, size, clip_rect, series_p_draw_line, series_p_draw_line_area,
       series_p_draw_smooth_line, series_p_draw_smooth_line_area, series_p_draw_symbol);
@@ -285,7 +288,6 @@ ret_t line_series_on_paint_self(widget_t* widget, canvas_t* c) {
   return_value_if_fail(series != NULL, RET_BAD_PARAMS);
 
   line_series_start_init_if_not_inited(widget);
-  series_p_reset_fifo(widget);
 
   if (series->base.display_mode == SERIES_DISPLAY_COVER) {
     return series_p_on_paint_self_cover(widget, c);
@@ -294,14 +296,20 @@ ret_t line_series_on_paint_self(widget_t* widget, canvas_t* c) {
   }
 }
 
-static const char* s_line_series_properties[] = {
-    SERIES_PROP_CAPACITY,     SERIES_PROP_UNIT_SIZE,       SERIES_PROP_COVERAGE,
-    SERIES_PROP_DISPLAY_MODE, SERIES_PROP_VALUE_ANIMATION, SERIES_PROP_TITLE,
-    SERIES_PROP_LINE_SHOW,    SERIES_PROP_LINE_SMOOTH,     SERIES_PROP_LINE_AREA_SHOW,
-    SERIES_PROP_SYMBOL_SIZE,  SERIES_PROP_SYMBOL_SHOW,     NULL};
+static const char* s_line_series_properties[] = {SERIES_PROP_FIFO,
+                                                 SERIES_PROP_COVERAGE,
+                                                 SERIES_PROP_DISPLAY_MODE,
+                                                 SERIES_PROP_VALUE_ANIMATION,
+                                                 SERIES_PROP_TITLE,
+                                                 SERIES_PROP_LINE_SHOW,
+                                                 SERIES_PROP_LINE_SMOOTH,
+                                                 SERIES_PROP_LINE_AREA_SHOW,
+                                                 SERIES_PROP_SYMBOL_SIZE,
+                                                 SERIES_PROP_SYMBOL_SHOW,
+                                                 NULL};
 
 static const series_draw_data_info_t s_line_series_draw_data_info = {
-    .size = sizeof(series_p_draw_data_t),
+    .unit_size = sizeof(series_data_draw_normal_t),
     .compare_in_axis1 = series_p_draw_data_compare_x,
     .compare_in_axis2 = series_p_draw_data_compare_y,
     .min_axis1 = series_p_draw_data_min_x,
@@ -323,13 +331,12 @@ static const series_vtable_t s_line_series_internal_vtable = {
     .is_point_in = series_p_is_point_in,
     .index_of_point_in = series_p_index_of_point_in,
     .to_local = series_p_to_local,
-    .set = series_p_set_default,
+    .set = series_p_set,
     .on_paint = line_series_on_paint,
     .draw_data_info = &s_line_series_draw_data_info};
 
 TK_DECL_VTABLE(line_series) = {.size = sizeof(line_series_t),
                                .type = WIDGET_TYPE_LINE_SERIES,
-                               .enable_pool = TRUE,
                                .parent = TK_PARENT_VTABLE(series),
                                .clone_properties = s_line_series_properties,
                                .persistent_properties = s_line_series_properties,
@@ -354,8 +361,14 @@ widget_t* line_series_create_internal(widget_t* parent, xy_t x, xy_t y, wh_t w, 
 }
 
 widget_t* line_series_create(widget_t* parent, xy_t x, xy_t y, wh_t w, wh_t h) {
-  return line_series_create_internal(parent, x, y, w, h, TK_REF_VTABLE(line_series),
-                                     &s_line_series_internal_vtable);
+  widget_t* widget = line_series_create_internal(parent, x, y, w, h, TK_REF_VTABLE(line_series),
+                                                 &s_line_series_internal_vtable);
+  return_value_if_fail(widget != NULL, NULL);
+
+  object_t* fifo = series_fifo_default_create(10, sizeof(float_t));
+  series_p_set_fifo(widget, fifo);
+
+  return widget;
 }
 
 widget_t* line_series_cast(widget_t* widget) {